Update verify version check to spam less and settable by camera
First, this change allows toggling the version checks by camera.
This is nicer for multi-camera setups in cases where you may be
experimenting with several cameras, or have a few cameras that are
not as important as others.

Second, this only allows showing the version error once per
camera, showing the 'no coprocessors' errors once, annd the
disconnected error message once per case of connection issues.